View for comment permissions
Comments can now be tagged to a writer. If a writer is logged in a posts a comment, the comment will be linked to his/her account, regardless of the name he/she chooses to comment with. If the writer leaves the space blank, then his/her account name will be used instead of anonymous.

A bar which includes an edit and a delete glyphicon is also added to the comment boxes. They will only appear for comments which the user have ownership of, i.e. for writers who have commented while logged in. The delete button is fully functional with a confirmation. The edit button is not yet implemented.hi

While in the UI the buttons are simply hidden, the server do have guards to prevent rogue POST requests without the proper credentials to modify a comment.

Signed-off-by: Shen Yichen <2007.yichen@gmail.com>